requirements.txt
asgiref
attrs
bleach
cachetools
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-cors-headers
django-debug-toolbar
django-environ
django-extra-settings
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-nested-routers
drf-spectacular
drf-standardized-errors
facebook-sdk
future
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
inflection
jsonfield
jsonschema
jsonschema-specifications
oauthlib
packaging
protobuf
psycopg2
pyasn1
pyasn1-modules
pycparser
PyJWT
pyparsing
python-dotenv
python-magic-bin
python-twitter
python3-openid
pytz
PyYAML
referencing
requests
requests-oauthlib
rpds-py
rsa
ruff
six
social-auth-app-django
social-auth-core
sqlparse
tzdata
uritemplate
urllib3
uuid
webencodings
